In Rev 1.130 of GraphicsOut.cxx by drwr, the query for render-to-texture support is no longer made to the gsg directly at line 368, but is instead routed through an internal GraphicsOut.get_supports_render_texture() which defaults to false. In the wgl path, the derived wglGraphicsBuffer supplies its own get_supports_render_texture() that queries the gsg. The DirectX equivalent classes do not supply an overriding get_supports_render_texture() and so fails everytime when trying to render_to_texture.
Hence, why distortion the sample is failing under Dx9.

This set of fixes patches in the requisite get_supports_render_texture() for wdxGraphicsBuffer9 and fixes a few typos between depth and color texture usages. It also fixes a depth buffer bug with render-to-textures where the depth check appeared to be missing.

Thanks to Jonah (11thpenguin) for these fixes.